pyproject.toml
[project]
name = "test-backend-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "httpx<0.28.0",
    "openai==1.51.2",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "uvicorn[standard]==0.30.6",
]
